Revert "Filter done and some work on report"
This reverts commit bac41cad9f3e17296f8e995f220ef02a3926bd08.

+++ BookingSystem/BookingSystem/Filter.h
@@ -14,7 +14,7 @@ namespace lic {
 		std::string _text;
 		/* Minimum rating of the service
 		*/
-		float _rating;
+		int _rating;
 		/* Maximum age limit of the games/services.
 		*/
 		int _ageRestriction;
